move some methods from utils to logic folder
