+++ generators/languages/entity-files.mjs
@@ -18,6 +18,7 @@
  */
 import _ from 'lodash';
 import utils from '../utils.cjs';
+import { getEnumInfo } from '../entity/logic/index.mjs';
 
 const { startCase } = _;
 
@@ -69,7 +70,7 @@ export function writeEntityFiles() {
                   this.writeFiles({
                     sections: enumClientI18nFiles,
                     context: {
-                      ...utils.getEnumInfo(field, entity.clientRootFolder),
+                      ...getEnumInfo(field, entity.clientRootFolder),
                       lang,
                       frontendAppName,
                       packageName,
