+++ generators/server/logic/index.mjs
@@ -17,6 +17,4 @@
  * limitations under the License.
  */
 // eslint-disable-next-line import/prefer-default-export
-export { default as entityIsAuthority } from './entity/authority-checker.mjs';
-export { default as entityIsUser } from './entity/user-checker.mjs';
-export { default as getEnumInfo } from './enum.mjs';
+export { default as javadoc } from './text-formatting.mjs';
